+++ src/test/java/tfw/immutable/ila/doubleila/DoubleIlaFromCastLongIlaTest.java
@@ -24,7 +24,7 @@ class DoubleIlaFromCastLongIlaTest {
         LongIla ila = LongIlaFromArray.create(array);
         DoubleIla targetIla = DoubleIlaFromArray.create(target);
         DoubleIla actualIla = DoubleIlaFromCastLongIla.create(ila);
-        final double epsilon = (double) 0.0;
+        final double epsilon = 0.0;
         DoubleIlaCheck.checkAll(
                 targetIla,
                 actualIla,
