+++ src/test/java/tfw/immutable/ila/doubleila/DoubleIlaNegateTest.java
@@ -17,12 +17,12 @@ class DoubleIlaNegateTest {
         final double[] target = new double[length];
         for (int ii = 0; ii < array.length; ++ii) {
             array[ii] = random.nextDouble();
-            target[ii] = (double) -array[ii];
+            target[ii] = -array[ii];
         }
         DoubleIla ila = DoubleIlaFromArray.create(array);
         DoubleIla targetIla = DoubleIlaFromArray.create(target);
         DoubleIla actualIla = DoubleIlaNegate.create(ila);
-        final double epsilon = (double) 0.0;
+        final double epsilon = 0.0;
         DoubleIlaCheck.checkAll(
                 targetIla,
                 actualIla,
